+++ integration/test/register.go
@@ -2,7 +2,6 @@ package test
 
 import (
 	"bytes"
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -103,7 +102,7 @@ func testMailRegisterOK() error {
 	expect := okResponse{
 		Status: "ok",
 		Body: map[string]interface{}{
-			"mail": "test@gmail.com",
+			"mail":     "test@gmail.com",
 			"verified": true,
 		},
 	}
@@ -197,7 +196,7 @@ func testGetUserInfo() error {
 	if err != nil {
 		return fmt.Errorf("Expect no error, but %v", err)
 	}
-	req.Header.Set("Authorization", "jwt "+base64.RawStdEncoding.EncodeToString([]byte(accessToken)))
+	req.Header.Set("Authorization", "jwt "+accessToken)
 	client := http.Client{}
 	resp, err := client.Do(req)
 
@@ -251,7 +250,7 @@ func testUpdateUserInfo() error {
 		return fmt.Errorf("Expect no error, but %v", err)
 	}
 
-	req.Header.Set("Authorization", "jwt "+base64.RawStdEncoding.EncodeToString([]byte(accessToken)))
+	req.Header.Set("Authorization", "jwt "+accessToken)
 	req.Header.Set("Content-type", "application/json")
 	client := http.Client{}
 	resp, err := client.Do(req)